Defined notifications for silent newsletter
We have a scenario where we have existing newsletter subscription module but after each action of the user no notification is displayed and the navigation is left at halfway. So I have created few high priority notification tasks with screenshot and details how to solve the issue and how to be tested by the QA team.
Note: The template that I am using in all of the Task Definitions is created by me (of course after some google research of the most appropriate way to define a task).